Review changes: rename mocks to stubs, remove unused dependencies
