+++ packages/client/jest.config.js
@@ -13,5 +13,5 @@ module.exports = {
     "^.+\\.(ts|tsx)$": "ts-jest",
   },
   transformIgnorePatterns: ["/node_modules/", "\\.pnp\\.[^\\/]+$"],
-  testPathIgnorePatterns: ["node_modules", "__tests__/__mocks__"],
+  testPathIgnorePatterns: ["node_modules", "__tests__/__stubs__"],
 };
