+++ packages/client/package.json
@@ -23,13 +23,12 @@
     "@slicemachine/core": "1.1.10-alpha.0",
     "axios": "^0.27.2",
     "mime": "^2.5.2",
+    "form-data": "^3.0.0",
     "fp-ts": "^2.11.5",
     "io-ts": "^2.2.16"
   },
   "devDependencies": {
     "@babel/core": "^7.15.8",
-    "@babel/preset-env": "^7.15.8",
-    "@babel/preset-typescript": "^7.15.0",
     "@types/axios": "^0.14.0",
     "@types/jest": "^27.0.2",
     "@types/mime": "^2.0.3",
@@ -45,8 +44,6 @@
     "jest": "^27.2.4",
     "nock": "^13.1.3",
     "ts-jest": "^27.0.5",
-    "types-package-json": "^2.0.39",
-    "typescript": "^4.4.3",
-    "unionfs": "^4.4.0"
+    "typescript": "^4.4.3"
   }
 }
