+++ packages/core/package.json
@@ -20,7 +20,6 @@
   },
   "dependencies": {
     "@prismicio/types-internal": "^0.2.8",
-    "axios": "^0.26.0",
     "cookie": "^0.4.1",
     "fp-ts": "^2.11.5",
     "io-ts": "^2.2.16",
@@ -29,9 +28,6 @@
   },
   "devDependencies": {
     "@babel/core": "^7.15.8",
-    "@babel/preset-env": "^7.15.8",
-    "@babel/preset-typescript": "^7.15.0",
-    "@types/axios": "^0.14.0",
     "@types/cookie": "^0.4.1",
     "@types/jest": "^27.0.2",
     "@types/node": "^16.10.5",
@@ -45,10 +41,8 @@
     "eslint-plugin-jest": "^24.4.0",
     "jest": "^27.2.4",
     "memfs": "^3.3.0",
-    "nock": "^13.1.3",
     "ts-jest": "^27.0.5",
     "types-package-json": "^2.0.39",
-    "typescript": "^4.4.3",
-    "unionfs": "^4.4.0"
+    "typescript": "^4.4.3"
   }
 }
